+++ Automate/Properties/AssemblyInfo.cs
@@ -1,5 +1,5 @@
 using System.Reflection;
 
 [assembly: AssemblyTitle("Automate")]
-[assembly: AssemblyVersion("1.10.6")]
-[assembly: AssemblyFileVersion("1.10.6")]
+[assembly: AssemblyVersion("1.11.0")]
+[assembly: AssemblyFileVersion("1.11.0")]