+++ DataLayers/release-notes.md
@@ -1,7 +1,7 @@
 # Release notes
-## Upcoming release
-* Updated for the upcoming SMAPI 3.0.
+## 1.5
 * Added _machine processing_ layer (requires Automate 1.11+).
+* Updated for the upcoming SMAPI 3.0.
 * Improved translations. Thanks to Nanogamer7 (German)!
 
 ## 1.4.2